requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=6.0
numpy>=1.19.0
scipy>=1.5.0

# Code quality
black
isort
mypy

# Documentation
sphinx>=4.0
furo
sphinx-autodoc-typehints
sphinx-sitemap

# Building & Publishing
build
twine
ghp-import

# Optional CUDA backends
nvmath-python[cu12]
cupy-cuda12x>=13.0








